refac out of test
