+++ src/typescript/sdk/src/indexer-v2/queries/app/index.ts
@@ -1,3 +1,4 @@
 export * from "./home";
 export * from "./market";
 export * from "./pools";
+export * from "./dexscreener";
